+++ src/swap/defi/uni-v2-based/plugins/spookySwap.ts
@@ -56,8 +56,8 @@ export function makeSpookySwapPlugin(
 
     // Sanity check: Both wallets should be of the same chain.
     if (
-      fromWallet.currencyInfo.currencyCode !== 'fantom' ||
-      toWallet.currencyInfo.currencyCode !== 'fantom'
+      fromWallet.currencyInfo.pluginId !== 'fantom' ||
+      toWallet.currencyInfo.pluginId !== 'fantom'
     )
       throw new SwapCurrencyError(swapInfo, request)
 
@@ -104,14 +104,17 @@ export function makeSpookySwapPlugin(
 
     const fromAddress = (await fromWallet.getReceiveAddress()).publicAddress
     // toEdgeUnsignedTxs
-    const edgeSpendInfos = swapTxs.map(swapTx => {
+    const edgeSpendInfos = swapTxs.map((swapTx, i) => {
       // Convert to our spendInfo
       const edgeSpendInfo: EdgeSpendInfo = {
         currencyCode: request.fromCurrencyCode, // what is being sent out, only if token. Blank if not token
         spendTargets: [
           {
             memo: swapTx.data,
-            nativeAmount: swapTx.value != null ? swapTx.value.toString() : '0', // biggy/number string integer
+            nativeAmount:
+              swapTxs.length === 2 && i === 0
+                ? '0' // approval transactions don't have a value
+                : amountToSwap,
             publicAddress: swapTx.to
           }
         ],